pyproject.toml
[tool.poetry]
package-mode = false
name = "basic-autogen-tavily-tool"
version = "0.1.0"
description = "An application showing how to create an AutoGen tool using Tavily Search as an example"
authors = ["Lennex Zinyando <brizdigital@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
autogen = "^0.3.0"
pydantic = "^2.9.1"
tavily-python = "^0.5.0"
python-dotenv = "^1.0.1"
groq = "^0.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
